rename show page to detail page
