+++ mobile/routes/Navigator.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 import { NavigationContainer } from "@react-navigation/native";
-import { Landing, List, Show } from "../app/screens";
+import { Landing, List, Detail } from "../app/screens";
 
 const { Navigator, Screen } = createStackNavigator();
 
 const LandingNavigator = () => {
   return (
-    <Navigator initialRouteName="Show">
+    <Navigator initialRouteName="Landing">
       <Screen
         name="Landing"
         component={Landing}
@@ -16,7 +16,7 @@ const LandingNavigator = () => {
         }}
       />
       <Screen name="List" component={List} />
-      <Screen name="Show" component={Show} />
+      <Screen name="Detail" component={Detail} />
     </Navigator>
   );
 };
